all tests passing except instance methods
