+++ app/models/event.rb
@@ -1,9 +1,9 @@
 class Event < ActiveRecord::Base
   has_many :payloads
-  
+
   validates :name, presence: true
 
   def self.most_frequent_event_name
-    self.group(:name).order("count_name desc").count("name")
+    self.joins(:payloads).group("events.name").order(count: :desc).count
   end
 end
