+++ app/models/request_type.rb
@@ -2,10 +2,9 @@ require 'pry'
 class RequestType < ActiveRecord::Base
   has_many :payloads
 
-  validates :verb, presence: true, uniqueness: true
+  validates :verb, presence: true
 
   def self.list_distinct_verbs
     self.distinct(:verb).pluck(:verb)
   end
-
 end
