+++ test/models/payload_test.rb
@@ -63,55 +63,4 @@ class PayloadTest < Minitest::Test
     assert_equal "GET", Payload.most_frequent_request_type
   end
 
-  def test_returns_list_of_all_verbs_used_ever
-    skip
-    setup_1
-    assert_equal "GET", "POST" , Payload.method
-  end
-
-  def test_returns_max_response_time_for_given_url
-    client_url = "http://jumpstartlab.com"
-    setup_1
-    time = Payload.max_response_time_given_url(client_url)
-    assert_equal 30, time
-  end
-
-  def test_returns_min_response_time_for_given_url
-    client_url = "http://jumpstartlab.com/jumps"
-    setup_1
-    time = Payload.min_response_time_given_url(client_url)
-    assert_equal 40, time
-  end
-
-  def test_returns_all_response_times_for_given_url_sorted_by_length
-    client_url = "http://jumpstartlab.com"
-    setup_1
-    times = Payload.all_response_times_given_url(client_url)
-    assert_equal [30, 20], times
-    assert times.first > times.last
-  end
-
-  def test_return_average_response_time_for_given_url
-    client_url = "http://jumpstartlab.com"
-    setup_1
-    time = Payload.average_response_time_given_url(client_url)
-    assert_equal 25.0, time
-  end
-
-  def test_HTTP_verbs_used_to_acquire_this_url_returns_array
-    skip
-    client_url = "http://jumpstartlab.com"
-    setup_1
-    verbs = Payload.verbs_given_url(client_url)
-    assert_equal ["POST", "GET"], verbs
-  end
-
-  def test_three_most_popular_refers
-    skip
-    client_url = "http://jumpstartlab.com"
-    setup_1
-    referrers = Payload.three_most_popular_referrers(client_url)
-    assert_equal ["http://google.com", "http://turing.io"], referrers
-  end
-
 end
